ai_service/requirements.txt
fastapi
uvicorn[standard]
pydantic
openai
celery[redis]
python-dotenv
httpx # openai SDK >1.0 需要
eventlet